+++ financemanagerapp/changelog.md
@@ -2,4 +2,4 @@
 
 ## 1.3.0
 
-- 
+- Metadata file (BACKUP-META) is generated and included in backups. This file contains: version of the application, date of generation, number of files included, list of files included, checksum of each file